deployment/huggingface/requirements.txt

flask
flask-cors
torch
numpy
python-chess
